Fix Travis CI builds
The code to checkout the project from Git and to merge with master is not
necessary as that is done automatically by Travis CI in "travis-ci/pr"
checks. Furthermore, the original code would fail with "reference is not
a tree" for PRs. So fix and simplify the code, but still make HEAD point
to a branch, and have an upstream configured fopr the branch, as
otherwise test would not run.